tests/requirements-dev.txt
pytest==7.4.3
python-dotenv==1.0.0
supabase==2.0.3
pytest-cov==4.1.0 